Administration: - modification des commentaires OK - remaniement des paramètres :   - suppression des paramètres inutiles   - paramètres pour les commentaires, avec test de connec à la base de données   - paramètres pour le panorama ajax : temps entre photos et fading on/off - renomage d'une galerie => répercution sur la base - suppression d'une galerie => suppression des commentaires associés - correction d'un bug dans l'outil renomage : il ne renomait qu'une photo à la fois - diverses correction de typo
CSS:
- bugs divers sous IE
- la flèche du menu "vous êtes ici" quand on consulte les galeries est OK sous IE
- les commentaires sont css-isés

Partie client:
- bug png transparent: à moitié résolu en faisant un désentrelacement
- pas de génération de preview si la taille de la photo originale est plus petite


git-svn-id: file:///home/nicolas/svn/perso/luxbum/trunk@59 795f2d49-7102-0410-9315-962b31f5c57f

+++ _fonctions/class/files.php
@@ -68,30 +68,35 @@ class files {
     * Suppression récursive d'un répertoire (rm -rf)
     */
    function delTree ($dir)  {
-      $current_dir = opendir ($dir);
-      while ($entryname = readdir ($current_dir)) {
-         if (is_dir ($dir.'/'.$entryname) && ($entryname != '.' && $entryname!='..')) {
-            if (!files::deltree ($dir.'/'.$entryname)) {
-               return false;
+      if ($current_dir = opendir ($dir)) {
+         while ($entryname = readdir ($current_dir)) {
+            if (is_dir ($dir.'/'.$entryname) && ($entryname != '.' && $entryname!='..')) {
+               if (!files::deltree ($dir.'/'.$entryname)) {
+                  return false;
+               }
+            }
+            else if ($entryname != '.' && $entryname!='..') {
+               if (!@unlink ($dir.'/'.$entryname)) {
+                  return false;
+               }
             }
          }
-         else if ($entryname != '.' && $entryname!='..') {
-            if (!@unlink ($dir.'/'.$entryname)) {
-               return false;
-            }
-         }
+         closedir ($current_dir);
+         return @rmdir ($dir);
       }
-      closedir ($current_dir);
-      return @rmdir ($dir);
+      return false;
    }
 
    /**
     *
     */
    function writeFile ($file, $content, $mode = 'wt') {
-      $fd = fopen ($file, 'wt');
-      fwrite ($fd, $content);
-      fclose ($fd);
+      if ($fd = fopen ($file, $mode)) {
+         fwrite ($fd, $content);
+         fclose ($fd);
+         return true;
+      }
+      return false;
    }
 
    /**
@@ -122,7 +127,9 @@ class files {
             mkdir ($path, 0777);
          }
          umask ($old_umask);
+         return true;
       }
+      return false;
    }
 
    /**
